+++ c4x4/src/functions/cookies.ts
@@ -1,0 +1,15 @@
+export default function getCookie(name: String) {
+  const value = `; ${document.cookie}`
+  const parts = value.split(`; ${name}=`)
+  if (parts.length === 2) {
+    return parts.pop().split(';').shift();
+  }
+}
+
+export function deleteCookie(name: string) {
+  document.cookie = `${name}="; expires=Thu, 01 Jan 1970 00:00:01 GMT;"`;
+}
+
+export function setCookie(name: string, value: string) {
+  document.cookie = `${name}=${value}; SameSite=Strict`
+}