+++ src/tests/test_x4/test_dataset/test_habitats.py
@@ -1,3 +1,5 @@
+import json
+
 import pytest
 from rest_framework import status
 
@@ -9,7 +11,7 @@ class TestStationHabitats:
     def test_post(self, authed_client, create_habitat_module, create_station):
         response = authed_client.post(
             "/game/1/stations/1/habitats/",
-            {"data": [{"module_id": 1, "count": 5}]},
+            json.dumps({"data": [{"module_id": 1, "count": 5}]}),
             content_type="application/json",
         )
         assert response.status_code == status.HTTP_201_CREATED
@@ -61,7 +63,9 @@ class TestStationHabitatsView:
         }
 
     def test_get_404_on_not_my_station(
-        self, authed_client, create_user_2_habitat
+        self,
+        authed_client,
+        create_user_2_habitat,
     ):
         response = authed_client.get("/game/1/stations/1/habitats/1/")
         assert response.status_code == status.HTTP_404_NOT_FOUND