+++ src/x4companion/x4/models.py
@@ -4,6 +4,26 @@ from django.contrib.auth.models import User
 from django.db import models
 
 
+class Dataset(models.Model):
+    """Links a set of resources to be loaded into the app.
+
+    This model is referenced by template classes like HabitatModule or
+    FactoryModule. A Dataset is used to present a filtered view of templates,
+    so that only the content that is relevant to a SaveGame is accessible.
+
+    Attributes:
+        name (str): The name of the Dataset.
+
+    """
+
+    name = models.CharField(
+        max_length=50, blank=False, null=False, unique=True
+    )
+
+    def __str__(self) -> str:
+        return f"Dataset {self.name}"
+
+
 class SaveGame(models.Model):
     """The overarching save game that's being modeled.
 
@@ -15,13 +35,39 @@ class SaveGame(models.Model):
 
     name = models.CharField(max_length=50, blank=False, null=False)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
+    dataset = models.ForeignKey(Dataset, on_delete=models.CASCADE)
 
     def __str__(self) -> str:
         return f"SaveGame {self.name}"
 
 
-class Sector(models.Model):
+class SectorTemplate(models.Model):
     """Represents an in game sector.
+
+    Attributes:
+        name (str): The name of the sector.
+        dataset (Dataset): The Dataset this sector belongs to.
+
+    """
+
+    name = models.CharField(
+        max_length=50, unique=True, blank=False, null=False
+    )
+    dataset = models.ForeignKey(Dataset, on_delete=models.CASCADE)
+
+    class Meta:
+        constraints = [
+            models.UniqueConstraint(
+                fields=["name", "dataset"], name="No duplicate sectors"
+            )
+        ]
+
+    def __str__(self) -> str:
+        return f"Sector {self.name}"
+
+
+class Sector(models.Model):
+    """Represents an instance of a sector in a users save game.
 
     Attributes:
         name (str): The name of the sector.
@@ -29,20 +75,19 @@ class Sector(models.Model):
 
     """
 
-    name = models.CharField(
-        max_length=50, unique=True, blank=False, null=False
-    )
+    template = models.ForeignKey(SectorTemplate, on_delete=models.CASCADE)
     game = models.ForeignKey(SaveGame, on_delete=models.CASCADE)
 
     class Meta:
         constraints = [
             models.UniqueConstraint(
-                fields=["name", "game"], name="No duplicate sectors"
+                fields=["template", "game"],
+                name="No duplicate sector templates",
             )
         ]
 
     def __str__(self) -> str:
-        return f"Sector {self.name}"
+        return f"Sector {self.template.name}"
 
 
 class Station(models.Model):
@@ -78,26 +123,6 @@ class Station(models.Model):
             [mod.count * mod.module.capacity for mod in modules]
         )
         self.save()
-
-
-class Dataset(models.Model):
-    """Links a set of resources to be loaded into the app.
-
-    This model is referenced by template classes like HabitatModule or
-    FactoryModule. A Dataset is used to present a filtered view of templates,
-    so that only the content that is relevant to a SaveGame is accessible.
-
-    Attributes:
-        name (str): The name of the Dataset.
-
-    """
-
-    name = models.CharField(
-        max_length=50, blank=False, null=False, unique=True
-    )
-
-    def __str__(self) -> str:
-        return f"Dataset {self.name}"
 
 
 class HabitatModule(models.Model):