+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[0.1.0]
+### Added
+- Added Sector templates. These map to datasets, this table contains all sectors available in a dataset.
+- Added React frontend with material UI. This includes:
+  - Sign in and authentication.
+  - Save management.
+  - basic page layout for all future pages.
+- Added API endpoints for Sector Templates.
+
+### Changed
+- Sectors now map to sector templates, previously each sector was unique within a save game.
+- Save games must be associated to a dataset.
+
 ## [0.0.1] - 2024-12-20
 
 ### Added